Perform NH routine in CUDA
Now COM velocities as well as the kinetic energy sums for NH routine
are performed in CUDA kernel. This update enhances the performance
two to four times, making the performance to be about 40 % slower than
Dual Langevin, and a bit more than two times slower than Langevin
in case of nonpolarizable simulation.