Modify package name to drudeTGNH
This commit changes the package naming into drudetgnh plugin, so that
it can be distinguished from OpenMM7.5's updated drude plugin
and remove any conflicts with the drudenosehooverintegrator in there.
Note that the updated default drude plugin does not support temperature grouping,
which is essential to simulate any system with high ffrequency modes such as improper dihedral.